% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_my_lm.R
\name{print.my_lm}
\alias{print.my_lm}
\title{Print Model Information for Custom Linear Regression}
\usage{
\method{print}{my_lm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{my_lm}}

\item{...}{Not used.}
}
\description{
Provides improved output for displaying the fitted model using
the custom \code{\link[=my_lm]{my_lm()}} function.
}
\examples{
model = my_lm(mpg ~ disp, data = mtcars)

# Explicit call of the print function
print(model)

# Implicit call of the print function
model

}
