% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_my_lm.R
\name{summary.my_lm}
\alias{summary.my_lm}
\alias{print.summary.my_lm}
\title{Compute Inference for Custom Linear Regression}
\usage{
\method{summary}{my_lm}(object, ...)

\method{print}{summary.my_lm}(x, ...)
}
\arguments{
\item{object}{An object with the class of \code{my_lm}.}

\item{...}{Additional options.}

\item{x}{An object with the class of \code{summary.my_lm}.}
}
\value{
An object of class \code{summary.my_lm} that contains:
\itemize{
\item \code{est_mat}: A \code{matrix} with the following \strong{named} columns...
\itemize{
\item \code{Estimate}: The parameter (beta) estimated values
\item \verb{Std. Err}: The standard error values for the parameter estimates.
\item \verb{t value}: The t-test value for the parameter estimate.
\item \verb{Pr(>|t|)}: The p-value for the t-test value on the parameter.
}
\item \code{call}: Information on how the \code{\link[=my_lm]{my_lm()}} function was called.
}
}
\description{
Calculates standard error, t-test values, and p-values for each parameter
estimate.
}
\examples{
# Compute the model
model = my_lm(mpg ~ disp, data = mtcars)

# Calculate the inference values
model_inference = summary(model)

# Display inference values
model_inference

}
\seealso{
\code{\link[=my_lm]{my_lm()}}, \code{\link[=print.summary.my_lm]{print.summary.my_lm()}}
}
